refactor(api): renames /types/index.ts to /types/shared.ts
changes file name for shared types/schemas in API
